refactor(tags): tag-colors と color-palette-picker を features/tags へ移動
src/lib を feature 非依存に近づけるため、Tags 固有の色定義/解決ロジックと
ColorPaletteMenuItems を features/tags/ 配下に集約。

- lib/tag-colors.ts → features/tags/lib/tag-colors.ts
- lib/components/ui/color-palette-picker.{tsx,stories.tsx} → features/tags/components/
- 外部参照は features/tags barrel 経由に統一、内部は相対 import に変更
- 仕様/UI/DB 変更なし
