+++ apps/product/src/features/calendar/components/tag-filter/components/FilterItem/FilterItemMenu.tsx
@@ -14,8 +14,8 @@ import {
 } from 'lucide-react';
 import { useTranslations } from 'next-intl';
 
-import { IconPickerDropdownItems, TagIcon } from '@/features/tags';
-import { ColorPaletteMenuItems } from '@/lib/components/ui/color-palette-picker';
+import type { TagColorName } from '@/features/tags';
+import { ColorPaletteMenuItems, IconPickerDropdownItems, TagIcon } from '@/features/tags';
 import {
   DropdownMenuContent,
   DropdownMenuItem,
@@ -24,7 +24,6 @@ import {
   DropdownMenuSubContent,
   DropdownMenuSubTrigger,
 } from '@/lib/components/ui/dropdown-menu';
-import type { TagColorName } from '@/lib/tag-colors';
 import { cn } from '@/lib/utils';
 
 /** タグのグループ情報 */
