+++ apps/product/src/features/review/components/review/TagBreakdownBar.tsx
@@ -1,8 +1,8 @@
 'use client';
 
+import type { TagColorName } from '@/features/tags';
 import { TagIcon } from '@/features/tags';
 import { ColonTagLabel } from '@/lib/components/ui/colon-tag-label';
-import type { TagColorName } from '@/lib/tag-colors';
 import { cn } from '@/lib/utils';
 
 interface TagSegment {
