+++ apps/product/src/features/tags/components/color-palette-picker.tsx
@@ -4,10 +4,10 @@ import { Check } from 'lucide-react';
 import { useTranslations } from 'next-intl';
 
 import { DropdownMenuItem } from '@/lib/components/ui/dropdown-menu';
-import { TAG_COLOR_MAP, TAG_COLOR_NAMES, resolveTagColor } from '@/lib/tag-colors';
 import { cn } from '@/lib/utils';
+import { TAG_COLOR_MAP, TAG_COLOR_NAMES, resolveTagColor } from '../lib/tag-colors';
 
-import type { TagColorName } from '@/lib/tag-colors';
+import type { TagColorName } from '../lib/tag-colors';
 
 /**
  * 翻訳済みの色表示名を取得する
