+++ apps/product/src/lib/components/shell/sidebar/SidebarSection.stories.tsx
@@ -10,9 +10,9 @@ import { fn } from 'storybook/test';
 
 import { ChevronRight, Moon, PanelLeft, Plus, Search } from 'lucide-react';
 
+import { getTagColorClasses } from '@/features/tags';
 import { Checkbox } from '@/lib/components/ui/checkbox';
 import { MiniCalendar } from '@/lib/components/ui/mini-calendar';
-import { getTagColorClasses } from '@/lib/tag-colors';
 
 import { TagIcon } from '@/features/tags';
